added partitioning, did 'some' refactoring
