+++ extras/tests/TestFolderOperations/TestFolderOperations.ino
@@ -28,20 +28,20 @@ void printFolderContents(Folder dir, int indentation = 0) {
   // Print directories
   for (Folder subdir : directories) {
     for (int i = 0; i < indentation; i++) {
-      printFormatted("  ");
+      printToSerialOrRS485("  ");
     }
-    printFormatted("[D] ");
-    printFormatted(subdir.getPath());
+    printToSerialOrRS485("[D] ");
+    printToSerialOrRS485(subdir.getPath());
     printFolderContents(subdir, indentation + 1);
   }
 
   // Print files
   for (UFile file : files) {
     for (int i = 0; i < indentation; i++) {
-      printFormatted("  ");
+      printToSerialOrRS485("  ");
     }
-    printFormatted("[F] ");
-    printFormatted(file.getPath());
+    printToSerialOrRS485("[F] ");
+    printToSerialOrRS485(file.getPath());
   }
 }
 
@@ -49,12 +49,12 @@ void printFolderContents(Folder dir, int indentation = 0) {
 bool testFolderCreation(Folder root) {
   Folder subfolder = root.createSubfolder("test_folder");
   if (subfolder.exists()) {
-    printFormatted("\n--- Test creating folder using root.createSubfolder ---");
-    printFormatted("Test creating folder using root.createSubfolder - Success");
+    printToSerialOrRS485("\n--- Test creating folder using root.createSubfolder ---");
+    printToSerialOrRS485("Test creating folder using root.createSubfolder - Success");
     subfolder.remove();
     return true;
   } else {
-    printFormatted("Test creating folder using root.createSubfolder - Failed. Error: " + String(getErrno()));
+    printToSerialOrRS485("Test creating folder using root.createSubfolder - Failed. Error: " + String(getErrno()));
     return false;
   }
 }
@@ -62,18 +62,18 @@ bool testFolderCreation(Folder root) {
 bool testFolderRenaming(Folder root) {
   Folder sourceFolder = root.createSubfolder("source_folder");
   if (sourceFolder.exists()) {
-    printFormatted("\n--- Test renaming folder ---");
-    printFormatted("Source folder name: " + String(sourceFolder.getPathAsString()));
+    printToSerialOrRS485("\n--- Test renaming folder ---");
+    printToSerialOrRS485("Source folder name: " + String(sourceFolder.getPathAsString()));
     if (sourceFolder.rename("renamed_folder")) {
-      printFormatted("Folder renamed to: " + String(sourceFolder.getPathAsString()));
+      printToSerialOrRS485("Folder renamed to: " + String(sourceFolder.getPathAsString()));
       sourceFolder.remove();
       return true;
     } else {
-      printFormatted("Folder renaming failed. Error: " + String(getErrno()));
+      printToSerialOrRS485("Folder renaming failed. Error: " + String(getErrno()));
       return false;
     }
   } else {
-    printFormatted("Test folder renaming - Failed. Error: " + String(getErrno()));
+    printToSerialOrRS485("Test folder renaming - Failed. Error: " + String(getErrno()));
     return false;
   }
 }
@@ -83,24 +83,24 @@ bool testCopyingFolder(Folder root) {
   Folder copyDestination = root.createSubfolder("copy_destination");
 
   if (sourceFolder.exists()) {
-    printFormatted("\n--- Test copying a folder ---");
-    printFormatted("Source folder name: " + String(sourceFolder.getPathAsString()));
-    printFormatted("Destination folder name: " + String(copyDestination.getPathAsString()));
+    printToSerialOrRS485("\n--- Test copying a folder ---");
+    printToSerialOrRS485("Source folder name: " + String(sourceFolder.getPathAsString()));
+    printToSerialOrRS485("Destination folder name: " + String(copyDestination.getPathAsString()));
 
    
 
     if (sourceFolder.copyTo(copyDestination, true)) {
-      printFormatted("Folder copied successfully!");
+      printToSerialOrRS485("Folder copied successfully!");
       sourceFolder.remove();
       copyDestination.remove();
       return true;
     } else {
-      printFormatted("Folder copying failed. Error: " + String(getErrno()));
+      printToSerialOrRS485("Folder copying failed. Error: " + String(getErrno()));
       sourceFolder.remove();
       return false;
     }
   } else {
-    printFormatted("Test copying a folder - Failed to create source folder. Error: " + String(getErrno()));
+    printToSerialOrRS485("Test copying a folder - Failed to create source folder. Error: " + String(getErrno()));
     return false;
   }
 }
@@ -112,20 +112,20 @@ bool testMovingFolder(Folder root) {
   Folder moveDestination = root.createSubfolder("move_destination");
 
   if (sourceFolderMove.exists()) {
-    printFormatted("\n--- Test moving a folder ---");
-    printFormatted("Source folder name: " + String(sourceFolderMove.getPathAsString()));
-    printFormatted("Destination folder name: " + String(moveDestination.getPathAsString()));
+    printToSerialOrRS485("\n--- Test moving a folder ---");
+    printToSerialOrRS485("Source folder name: " + String(sourceFolderMove.getPathAsString()));
+    printToSerialOrRS485("Destination folder name: " + String(moveDestination.getPathAsString()));
     if (sourceFolderMove.moveTo(moveDestination)) {
-      printFormatted("Folder moved successfully!");
+      printToSerialOrRS485("Folder moved successfully!");
       sourceFolderMove.remove();
       moveDestination.remove();
       return true;
     } else {
-      printFormatted("Folder moving failed. Error: " + String(getErrno()));
+      printToSerialOrRS485("Folder moving failed. Error: " + String(getErrno()));
       return false;
     }
   } else {
-    printFormatted("Test moving a folder - Failed to create source folder. Error: " + String(getErrno()));
+    printToSerialOrRS485("Test moving a folder - Failed to create source folder. Error: " + String(getErrno()));
     return false;
   }
 
@@ -139,17 +139,17 @@ void runTests(Arduino_UnifiedStorage * storage, String storageType) {
         Folder root = storage->getRootFolder();
 
 
-        printFormatted("========= Folder Tests =========");
+        printToSerialOrRS485("========= Folder Tests =========");
 
         testFolderCreation(root);
         testFolderRenaming(root);
         testCopyingFolder(root);
         testMovingFolder(root);
 
-        printFormatted("========= FS Contents after Folder Tests =========");
+        printToSerialOrRS485("========= FS Contents after Folder Tests =========");
         printFolderContents(root);
         storage->unmount();
-        printFormatted("");
+        printToSerialOrRS485("");
     }
 }
