+++ extras/tests/TestRepeatedFormatMount/TestRepeatedFormatMount.ino
@@ -20,33 +20,33 @@ InternalStorage internal = InternalStorage();
 #endif
 
 void runRepeatedMountTest(Arduino_UnifiedStorage * storage, String storageType, int n = 10){
-  printFormatted("Running repeated mount test for " + storageType); printFormatted("\n");
+  printToSerialOrRS485("Running repeated mount test for " + storageType); printToSerialOrRS485("\n");
 
   for(size_t i=0;i<n;i++){
     int mountResult = storage->begin();
 
-    printFormatted("Mounting drive"); printFormatted("\n");
+    printToSerialOrRS485("Mounting drive"); printToSerialOrRS485("\n");
     if(mountResult != 1) {
-      printFormatted(mountResult);
-      printFormatted(getErrno());
-      printFormatted("\n");
+      printToSerialOrRS485(mountResult);
+      printToSerialOrRS485(getErrno());
+      printToSerialOrRS485("\n");
     }
 
     Folder root = storage->getRootFolder();
     UFile file = root.createFile("file.txt", FileMode::WRITE);
     file.write("writing stuff to the file");
     file.changeMode(FileMode::READ);
-    printFormatted(file.readAsString()); printFormatted("\n");
+    printToSerialOrRS485(file.readAsString()); printToSerialOrRS485("\n");
     file.close();
     file.remove();
 
     int umountResult = storage->unmount();
     if(!umountResult) {
-      printFormatted("Unmounting drive"); printFormatted("\n");
-      printFormatted(umountResult);
-      printFormatted(getErrno());
+      printToSerialOrRS485("Unmounting drive"); printToSerialOrRS485("\n");
+      printToSerialOrRS485(umountResult);
+      printToSerialOrRS485(getErrno());
     } else {
-      printFormatted("Successfully unmounted"); printFormatted("\n");
+      printToSerialOrRS485("Successfully unmounted"); printToSerialOrRS485("\n");
     }
   }
 }
@@ -61,43 +61,43 @@ void setup(){
     #endif 
     
     #if defined(HAS_QSPI)
-    printFormatted("RUNNING FORMAT AND REPEATED MOUNT - QSPI Storage \n"); printFormatted("\n");
-    printFormatted("Formatting QSPI drive as LittleFS: " + String(internal.format(FS_LITTLEFS))); printFormatted("\n");
+    printToSerialOrRS485("RUNNING FORMAT AND REPEATED MOUNT - QSPI Storage \n"); printToSerialOrRS485("\n");
+    printToSerialOrRS485("Formatting QSPI drive as LittleFS: " + String(internal.format(FS_LITTLEFS))); printToSerialOrRS485("\n");
     runRepeatedMountTest(&internal, "QSPI");
-    printFormatted("Formatting QSPI drive as FAT32: " + String(internal.format(FS_FAT))); printFormatted("\n");
+    printToSerialOrRS485("Formatting QSPI drive as FAT32: " + String(internal.format(FS_FAT))); printToSerialOrRS485("\n");
     runRepeatedMountTest(&internal, "QSPI");
-    printFormatted("Formatting SD drive as LittleFS again: " + String(internal.format(FS_LITTLEFS))); printFormatted("\n");
+    printToSerialOrRS485("Formatting SD drive as LittleFS again: " + String(internal.format(FS_LITTLEFS))); printToSerialOrRS485("\n");
     runRepeatedMountTest(&internal, "QSPI");
-    printFormatted("Formatting SD drive as FAT32 again: " + String(internal.format(FS_FAT))); printFormatted("\n");
+    printToSerialOrRS485("Formatting SD drive as FAT32 again: " + String(internal.format(FS_FAT))); printToSerialOrRS485("\n");
     runRepeatedMountTest(&internal, "QSPI");
     #endif 
 
     // format storage as FAT32
     #if defined(HAS_USB)
-    printFormatted("RUNNING FORMAT AND REPEATED MOUNT - USB \n"); printFormatted("\n");
-    printFormatted("Formatting USB drive as LittleFS: " + String(usb.format(FS_FAT))); printFormatted("\n");
-    printFormatted("" + String(errno));
+    printToSerialOrRS485("RUNNING FORMAT AND REPEATED MOUNT - USB \n"); printToSerialOrRS485("\n");
+    printToSerialOrRS485("Formatting USB drive as LittleFS: " + String(usb.format(FS_FAT))); printToSerialOrRS485("\n");
+    printToSerialOrRS485("" + String(errno));
     runRepeatedMountTest(&usb, "USB");
-    printFormatted("Formatting USB drive as FAT32: " + String(usb.format(FS_FAT)));  printFormatted("\n");
+    printToSerialOrRS485("Formatting USB drive as FAT32: " + String(usb.format(FS_FAT)));  printToSerialOrRS485("\n");
     runRepeatedMountTest(&usb, "USB");
         
-    printFormatted("Formatting USB drive as LittleFS again: " + String(usb.format(FS_LITTLEFS))); printFormatted("\n");
+    printToSerialOrRS485("Formatting USB drive as LittleFS again: " + String(usb.format(FS_LITTLEFS))); printToSerialOrRS485("\n");
     runRepeatedMountTest(&usb, "USB");
-    printFormatted("Formatting USB drive as FAT32 again: " + String(usb.format(FS_FAT))); printFormatted("\n");
-    runRepeatedMountTest(&usb, "USB"); printFormatted("\n");
+    printToSerialOrRS485("Formatting USB drive as FAT32 again: " + String(usb.format(FS_FAT))); printToSerialOrRS485("\n");
+    runRepeatedMountTest(&usb, "USB"); printToSerialOrRS485("\n");
     #endif 
 
 
 
     #if defined(HAS_SD)
-    printFormatted("RUNNING FORMAT AND REPEATED MOUNT - SD Card \n");
-    printFormatted("Formatting SD drive as LittleFS: " + String(sd.format(FS_LITTLEFS)));
+    printToSerialOrRS485("RUNNING FORMAT AND REPEATED MOUNT - SD Card \n");
+    printToSerialOrRS485("Formatting SD drive as LittleFS: " + String(sd.format(FS_LITTLEFS)));
     runRepeatedMountTest(&&sd, "SD");
-    printFormatted("Formatting SD drive as FAT32: " + String(sd.format(FS_FAT)));
+    printToSerialOrRS485("Formatting SD drive as FAT32: " + String(sd.format(FS_FAT)));
     runRepeatedMountTest(&&sd, "SD");
-    printFormatted("Formatting SD drive as LittleFS again: " + String(sd.format(FS_LITTLEFS)));
+    printToSerialOrRS485("Formatting SD drive as LittleFS again: " + String(sd.format(FS_LITTLEFS)));
     runRepeatedMountTest(&&sd, "SD");
-    printFormatted("Formatting SD drive as FAT32 again: " + String(sd.format(FS_FAT)));
+    printToSerialOrRS485("Formatting SD drive as FAT32 again: " + String(sd.format(FS_FAT)));
     runRepeatedMountTest(&&sd, "SD");
     #endif
