+++ src/Arduino_UnifiedStorage.h
@@ -3,28 +3,13 @@
 #define UnifiedStorage_H
 
 
-#include "Folder.h"
-#include "UFile.h"
-#include "Output.h"
-
-#if defined(ARDUINO_PORTENTA_C33)
-#include "Types.h"
 #include "Arduino.h"
 #include "Arduino_POSIXStorage.h"
-#endif
+#include "Folder.h"
+#include "UFile.h"
+#include "Utils.h"
 
 
-
-    #if defined(ARDUINO_PORTENTA_C33)
-        #include "QSPIFlashBlockDevice.h"
-        #include <BlockDevice.h>
-        #include <MBRBlockDevice.h>
-        #include "FATFileSystem.h"
-    #elif defined(ARDUINO_PORTENTA_H7_M7) || defined(ARDUINO_PORTENTA_H7_M4) || defined(ARDUINO_OPTA)
-        #include "QSPIFBlockDevice.h"
-        #include "MBRBlockDevice.h"
-        #include "FATFileSystem.h"
-    #endif 
 
 
 /**
