+++ src/InternalStorage.cpp
@@ -1,18 +1,9 @@
 #include "Arduino_UnifiedStorage.h"
 
 InternalStorage::InternalStorage(){
-/*
-    #if defined(ARDUINO_PORTENTA_C33)
-        this -> setQSPIPartition(2);
-    #elif defined(ARDUINO_OPTA)
-        this -> setQSPIPartition(3);
-    #elif defined(ARDUINO_PORTENTA_H7_M7)
-        this -> setQSPIPartition(3);
-    #endif
-*/
     this -> setQSPIPartitionName("user");
-  
 }
+
 
 InternalStorage::InternalStorage(int partition, const char * name, FileSystems fs){
     this -> setQSPIPartition(partition);
@@ -25,39 +16,22 @@ bool InternalStorage::begin(FileSystems fs){
   return this -> begin();
 }
 
+bool InternalStorage::partition(std::vector<Partition> partitions){
+    Partitioning::partitionDrive(QSPIFBlockDeviceType::get_default_instance(), partitions);
+}
+
 bool InternalStorage::begin(){
-    #if defined(ARDUINO_PORTENTA_C33)
-        this -> blockDevice = BlockDevice::get_default_instance();
-        this -> userData = new MBRBlockDevice(this->blockDevice, this->partitionNumber);
+        this -> blockDevice = BlockDeviceType::get_default_instance();
+        this -> userData = new MBRBlockDeviceType(this->blockDevice, this->partitionNumber);
+
         if(this -> fileSystem == FS_FAT){
-            this -> userDataFileSystem = new FATFileSystem(this->partitionName);
+            this -> userDataFileSystem = new FATFileSystemType(this->partitionName);
         } else {
-            this -> userDataFileSystem = new LittleFileSystem(this->partitionName);
+            this -> userDataFileSystem = new LittleFileSystemType(this->partitionName);
         }
+
         int err = this -> userDataFileSystem -> mount(userData);
         return err == 0;
-    #elif defined(ARDUINO_PORTENTA_H7_M7) ||  defined(ARDUINO_OPTA) 
-        this -> blockDevice = QSPIFBlockDevice::get_default_instance();
-        this -> userData = new mbed::MBRBlockDevice(this->blockDevice, this->partitionNumber);
-        if(this -> fileSystem == FS_FAT){
-            
-            if(this -> userDataFileSystem != nullptr){
-                delete(this -> userDataFileSystem);
-            } 
-            this -> userDataFileSystem = new mbed::FATFileSystem(this->partitionName);
-        } else {
-
-            if(this -> userDataFileSystem != nullptr){
-                delete(this -> userDataFileSystem);
-            } 
-
-            this -> userDataFileSystem = new mbed::LittleFileSystem(this->partitionName);
-        }
-        int err = this -> userDataFileSystem -> mount(this -> userData);
-        return err == 0;
-    #else
-        return false; // Unsupported board
-    #endif
 }
 
 bool InternalStorage::unmount(){
@@ -82,38 +56,18 @@ bool InternalStorage::format(FileSystems fs){
     this -> unmount();
     this -> fileSystem = fs;
 
-
     if(fs == FS_FAT){
-        #if defined(ARDUINO_PORTENTA_C33)
-            this -> userDataFileSystem = new FATFileSystem(this->partitionName);
+            this -> userDataFileSystem = new FATFileSystemType(this->partitionName);
             return this -> userDataFileSystem -> reformat(this-> userData)  == 0;
-        #elif defined(ARDUINO_PORTENTA_H7_M7) ||  defined(ARDUINO_OPTA) 
-            this -> userDataFileSystem =  new mbed::FATFileSystem(this->partitionName);
-            return this -> userDataFileSystem -> reformat(this-> userData)  == 0;
-        #endif
     } if (fs == FS_LITTLEFS) {
-        #if defined(ARDUINO_PORTENTA_C33)
-            this -> userDataFileSystem = new LittleFileSystem(this->partitionName);
+            this -> userDataFileSystem =  new LittleFileSystemType(this->partitionName);
             return this -> userDataFileSystem -> reformat(this-> userData)  == 0;
-        #elif defined(ARDUINO_PORTENTA_H7_M7) ||  defined(ARDUINO_OPTA) 
-            this -> userDataFileSystem =  new mbed::LittleFileSystem(this->partitionName);
-            return this -> userDataFileSystem -> reformat(this-> userData)  == 0;
-        #endif
     }
 
     return false;
 }
 
-#if defined(ARDUINO_PORTENTA_C33)
-BlockDevice * InternalStorage::getBlockDevice(){
+
+BlockDeviceType * InternalStorage::getBlockDevice(){
     return this -> blockDevice;
 }
-#endif
-
-
-#if defined(ARDUINO_PORTENTA_H7_M7) || defined(ARDUINO_OPTA) 
-mbed::BlockDevice *  InternalStorage::getBlockDevice(){
-    return this -> blockDevice;
-}
- 
-#endif
